requirements.txt
aniso8601==6.0.0
arabic-reshaper==2.0.14
attrs==19.1.0
Click==7.0
configparser==3.7.4
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
Pillow==6.0.0
pyrsistent==0.15.1
python-bidi==0.4.0
pytz==2019.1
reportlab==3.5.21
six==1.12.0
Werkzeug==0.15.2
